requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==13.3.0
firebase==3.0.1
firebase-admin==5.2.0
flake8==4.0.1
google-api-core==2.5.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.2
kafka-python==2.0.2
Markdown==3.3.6
mccabe==0.6.1
msgpack==1.0.3
numpy==1.22.2
opencv-python-headless==4.5.5.62
packaging==21.3
Pillow==9.0.1
proto-plus==1.20.3
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.63.0
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
zipp==3.7.0
